+++ packages/shared/src/index.ts
@@ -157,6 +157,7 @@ export * from "./utils/speaker-segments.js";
 export * from "./utils/sprite-labels.js";
 export * from "./utils/managed-generation-parameters.js";
 export * from "./utils/conversation-presence.js";
+export * from "./utils/scene-presence.js";
 export * from "./utils/avatar-crop.js";
 export * from "./utils/persona-normalization.js";
 export * from "./utils/noodle-mentions.js";